Move timer services to separate folder
